+++ packages/nx-firebase/src/generators/application/files/readme.md__tmpl__
@@ -16,7 +16,7 @@ This Nx Firebase application was generated by [@simondotm/nx-firebase](https://g
 
 ## Generated dependencies
 
-Nx-Firebase will add `firebase-tools`, `firebase-admin` and `firebase-functions` to your workspace `package.json` at the `'latest'` version. You may wish to set these to a specific version.
+Nx-Firebase will add `firebase-tools`, `firebase-admin` and `firebase-functions` to your workspace if they do not already exist.
 
 ## Next Steps
 
@@ -32,5 +32,8 @@ See the plugin [README](https://github.com/simondotm/nx-firebase/blob/main/READM
 * `nx run <%= projectName %>:deploy` - Deploy this app to firebase
 * `nx run <%= projectName %>:serve` - Serve this app using the firebase emulator
 * `nx run <%= projectName %>:build` - Build all functions associated with this app
+* `nx run <%= projectName %>:test` - Test all functions associated with this app
+* `nx run <%= projectName %>:lint` - Lint all functions associated with this app
+
