+++ packages/nx-firebase/src/generators/function/function.ts
@@ -12,7 +12,7 @@ import { applicationGenerator as nodeApplicationGenerator } from '@nrwl/node'
 
 import { initGenerator } from '../init/init'
 import {
-  calculateFirebaseConfigName,
+  getFirebaseConfigFromProject,
   getProjectName,
   updateTsConfig,
 } from '../../utils'
@@ -41,19 +41,11 @@ export function normalizeOptions(
     )
   }
 
-  // use firebase.<project>.json if it exists, otherwise fall back to firebase.json
-  const firebaseConfigName = calculateFirebaseConfigName(tree, firebaseApp)
-  // let firebaseConfigName = `firebase.${firebaseApp}.json`
-  // if (!tree.exists(firebaseConfigName)) {
-  //   // console.log(`looking for ${firebaseConfigName} failed, using fallback`)
-  //   firebaseConfigName = `firebase.json`
-  // }
-
-  // if (!tree.exists(firebaseConfigName)) {
-  //   throw new Error(
-  //     `Could not find firebase config called '${firebaseConfigName}' in this workspace.`,
-  //   )
-  // }
+  // read the firebase config used by the parent app project
+  const firebaseConfigName = getFirebaseConfigFromProject(
+    tree,
+    firebaseAppProject,
+  )
 
   return {
     ...options,
